+++ R/helper_functions.R
@@ -36,7 +36,8 @@
 convert_water <- function(water) {
   nms <- slotNames(water)
   lst <- lapply(nms, function(nm) slot(water, nm))
-  as.data.frame(setNames(lst, nms))
+  as.data.frame(setNames(lst, nms)) %>%
+    select(where(~ any(!is.na(.))))
 }
 
 #' Apply `define_water` and output a dataframe
@@ -152,7 +153,7 @@ define_water_chain <- function(df, output_water = "defined_water") {
 #'  `fn_once` or `fn_chain` functions. Note, parallel processing is best used when your code block takes more than a minute to run,
 #'  shorter run times will not benefit from parallel processing.
 #'
-#' @param df a data frame containing a column, defined_water, which has already been computed using \code{\link{define_water}}
+#' @param df a data frame containing a water class column, which has already been computed using \code{\link{define_water_chain}}
 #' @param input_water name of the column of water class data to be used as the input for this function. Default is "defined_water".
 #'
 #' @seealso \code{\link{balance_ions}}
@@ -203,7 +204,7 @@ balance_ions_once <- function(df, input_water = "defined_water") {
 #'  `fn_once` or `fn_chain` functions. Note, parallel processing is best used when your code block takes more than a minute to run,
 #'  shorter run times will not benefit from parallel processing.
 #'
-#' @param df a data frame containing a column, defined_water, which has already been computed using \code{\link{define_water}}
+#' @param df a data frame containing a water class column, which has already been computed using \code{\link{define_water_chain}}
 #' @param input_water name of the column of water class data to be used as the input for this function. Default is "defined_water".
 #' @param output_water name of the output column storing updated parameters with the class, water. Default is "balanced_water".
 #'
@@ -264,8 +265,8 @@ balance_ions_chain <- function(df, input_water = "defined_water", output_water =
 #'  `fn_once` or `fn_chain` functions. Note, parallel processing is best used when your code block takes more than a minute to run,
 #'  shorter run times will not benefit from parallel processing.
 #'
-#' @param df a data frame containing a column, defined_water, which has already
-#' been computed using \code{\link{define_water}} or \code{\link{balance_ions}}. The df may include columns named for the chemical(s) being dosed.
+#' @param df a data frame containing a water class column, which has already been computed using
+#' \code{\link{define_water_chain}}. The df may include columns named for the chemical(s) being dosed.
 #' @param input_water name of the column of water class data to be used as the input for this function. Default is "defined_water".
 #' @param hcl Hydrochloric acid: HCl -> H + Cl
 #' @param h2so4 Sulfuric acid: H2SO4 -> 2H + SO4
@@ -317,15 +318,11 @@ balance_ions_chain <- function(df, input_water = "defined_water", output_water =
 #'
 #' @export
 
-chemdose_ph_once <- function(df, input_water = "defined_water", hcl = 0, h2so4 = 0, h3po4 = 0, naoh = 0,
+chemdose_ph_once <- function(df, input_water = "defined_water",
+                             hcl = 0, h2so4 = 0, h3po4 = 0, naoh = 0,
                              na2co3 = 0, nahco3 = 0, caoh2 = 0, mgoh2 = 0,
                              cl2 = 0, naocl = 0, caocl2 = 0, co2 = 0,
                              alum = 0, fecl3 = 0, fe2so43 = 0, caco3 = 0) {
-
-  dosable_chems <- tibble(hcl, h2so4, h3po4, naoh,
-    na2co3, nahco3, caoh2, mgoh2,
-    cl2, naocl, caocl2, co2,
-    alum, fecl3, fe2so43, caco3)
 
   output <- df %>%
     chemdose_ph_chain(input_water = input_water, output_water = "dosed_chem_water",
@@ -357,8 +354,8 @@ chemdose_ph_once <- function(df, input_water = "defined_water", hcl = 0, h2so4 =
 #'  `fn_once` or `fn_chain` functions. Note, parallel processing is best used when your code block takes more than a minute to run,
 #'  shorter run times will not benefit from parallel processing.
 #'
-#' @param df a data frame containing a column, defined_water, which has already
-#' been computed using \code{\link{define_water}} or \code{\link{balance_ions}}. The df may include columns named for the chemical(s) being dosed.
+#' @param df a data frame containing a water class column, which has already been computed using
+#' \code{\link{define_water_chain}}. The df may include columns named for the chemical(s) being dosed.
 #' @param input_water name of the column of water class data to be used as the input for this function. Default is "defined_water".
 #' @param output_water name of the output column storing updated parameters with the class, water. Default is "dosed_chem_water".
 #' @param hcl Hydrochloric acid: HCl -> H + Cl
@@ -448,14 +445,12 @@ chemdose_ph_chain <- function(df, input_water = "defined_water", output_water = 
 
   } else if (nrow(chem_inputs_arg) > 1) {
 
-    chem_inputs_arg <- chem_inputs_arg %>%
-      mutate(ID = row_number())
     chem_doses <- chem_inputs_col %>%
-      left_join(chem_inputs_arg, by = "ID")
+      cross_join(chem_inputs_arg)
     chem2 <- dosable_chems %>%
       subset(select = !names(dosable_chems) %in% names(chem_doses)) %>%
-      mutate(ID = row_number()) %>%
-      left_join(chem_doses, by = "ID")
+      unique() %>%
+      cross_join(chem_doses)
   }
 
   output <- df %>%
@@ -485,7 +480,6 @@ chemdose_ph_chain <- function(df, input_water = "defined_water", output_water = 
 
 }
 
-
 #' Apply `solvedose_ph` to a dataframe and create a new column with numeric dose
 #'
 #' This function allows \code{\link{solvedose_ph}} to be added to a piped data frame.
@@ -504,8 +498,8 @@ chemdose_ph_chain <- function(df, input_water = "defined_water", output_water = 
 #'  `fn_once` or `fn_chain` functions. Note, parallel processing is best used when your code block takes more than a minute to run,
 #'  shorter run times will not benefit from parallel processing.
 #'
-#' @param df a data frame containing a column, defined_water, which has already
-#' been computed using \code{\link{define_water}}, and a column named for each of the chemicals being dosed
+#' @param df a data frame containing a water class column, which has already been computed using
+#' \code{\link{define_water_chain}}. The df may include a column with names for each of the chemicals being dosed.
 #' @param input_water name of the column of water class data to be used as the input. Default is "defined_water".
 #' @param output_column name of the output column storing doses in mg/L. Default is "dose_required".
 #' @param target_ph set a goal for pH using the function argument or a data frame column
@@ -605,8 +599,8 @@ solvedose_ph_once <- function(df, input_water = "defined_water", output_column =
 #'  `fn_once` or `fn_chain` functions. Note, parallel processing is best used when your code block takes more than a minute to run,
 #'  shorter run times will not benefit from parallel processing.
 #'
-#' @param df a data frame containing a column, defined_water, which has already
-#' been computed using \code{\link{define_water}}, and a column named for each of the chemicals being dosed
+#' @param df a data frame containing a water class column, which has already been computed using
+#' \code{\link{define_water_chain}}. The df may include a column with names for each of the chemicals being dosed.
 #' @param input_water name of the column of water class data to be used as the input. Default is "defined_water".
 #' @param output_column name of the output column storing doses in mg/L. Default is "dose_required".
 #' @param target_alk set a goal for alkalinity using the function argument or a data frame column
@@ -686,7 +680,6 @@ solvedose_alk_once <- function(df, input_water = "defined_water", output_column 
       target_alk = target_alk),
     solvedose_alk)) %>%
     mutate(!!output_column := as.numeric(dose)) %>%
-
     select(-dose)
 }
 
@@ -707,8 +700,8 @@ solvedose_alk_once <- function(df, input_water = "defined_water", output_column 
 #'  `fn_once` or `fn_chain` functions. Note, parallel processing is best used when your code block takes more than a minute to run,
 #'  shorter run times will not benefit from parallel processing.
 #'
-#' @param df a data frame containing a column, defined_water, which has already
-#' been computed using \code{\link{define_water}}, and a column named for each of the chemicals being dosed
+#' @param df a data frame containing a water class column, which has already been computed using
+#' \code{\link{define_water_chain}}
 #' @param waters List of column names containing a water class to be blended
 #' @param ratios List of column names or vector of blend ratios in the same order as waters. (Blend ratios must sum to 1)
 #'
@@ -799,8 +792,8 @@ blend_waters_once <- function(df, waters, ratios) {
 #'  `fn_once` or `fn_chain` functions. Note, parallel processing is best used when your code block takes more than a minute to run,
 #'  shorter run times will not benefit from parallel processing.
 #'
-#' @param df a data frame containing a column, defined_water, which has already
-#' been computed using \code{\link{define_water}}, and a column named for each of the chemicals being dosed
+#' @param df a data frame containing a water class column, which has already
+#' been computed using \code{\link{define_water_chain}},
 #' @param waters List of column names containing a water class to be blended
 #' @param ratios List of column names or vector of blend ratios in the same order as waters. (Blend ratios must sum to 1)
 #' @param output_water name of output column storing updated parameters with the class, water. Default is "blended_water".
@@ -865,7 +858,7 @@ blend_waters_chain <- function(df, waters, ratios, output_water = "blended_water
 #' This function plucks a selected parameter from a column of `water` class objects.
 #' To view multiple parameters, please use one of the "fn_once" functions or \code{\link{convert_water}}.
 #'
-#' @param df a data frame containing a column, defined_water, which has already
+#' @param df a data frame containing a water class column, which has already
 #' been computed using \code{\link{define_water}}
 #' @param input_water name of the column of water class data to be used as the input for this function.
 #' @param parameter water class attribute to view outside the water column
@@ -957,7 +950,6 @@ pluck_water <- function(df, input_water = "defined_water", parameter, output_col
 
 }
 
-
 #' Apply `dissolve_pb` to a dataframe and create a new column with numeric dose
 #'
 #' This function allows \code{\link{dissolve_pb}} to be added to a piped data frame.
@@ -974,8 +966,8 @@ pluck_water <- function(df, input_water = "defined_water", parameter, output_col
 #'  `fn_once` or `fn_chain` functions. Note, parallel processing is best used when your code block takes more than a minute to run,
 #'  shorter run times will not benefit from parallel processing.
 #'
-#' @param df a data frame containing a column, defined_water, which has already
-#' been computed using \code{\link{define_water}}, and a column named for each of the chemicals being dosed
+#' @param df a data frame containing a water class column, which has already been computed using
+#' \code{\link{define_water_chain}}
 #' @param input_water name of the column of water class data to be used as the input. Default is "defined_water".
 #' @param output_col_solid name of the output column storing the controlling lead solid. Default is "controlling_solid".
 #' @param output_col_result name of the output column storing dissolved lead in M. Default is "pb".
@@ -1044,6 +1036,212 @@ dissolve_pb_once <- function(df, input_water = "defined_water", output_col_solid
       rename(!!output_col_result := tot_dissolved_pb,
         !!output_col_solid := controlling_solid)
   }
+}
+
+#' Apply `chemdose_toc` function and output a dataframe
+#'
+#' This function allows \code{\link{chemdose_toc}} to be added to a piped data frame.
+#' Its output is a data frame with updated TOC, DOC, and UV254.
+#'
+#' The data input comes from a `water` class column, as initialized in \code{\link{define_water}} or \code{\link{balance_ions}}.
+#'
+#' If the input data frame has a column(s) name matching a valid coagulant(s), the function will dose that coagulant(s). Note:
+#' The function can only dose a coagulant as either a column or from the function arguments, not both.
+#'
+#' The column names must match the coagulant names as displayed in \code{\link{chemdose_toc}}.
+#' To see which coagulants can be passed into the function, see \code{\link{chemdose_toc}}.
+#'
+#' tidywater functions cannot be added after this function because they require a `water` class input.
+#'
+#'  For large datasets, using `fn_once` or `fn_chain` may take many minutes to run. These types of functions use the furrr package
+#'  for the option to use parallel processing and speed things up. To initialize parallel processing, use
+#'  `plan(multisession)` or `plan(multicore)` (depending on your operating system) prior to your piped code with the
+#'  `fn_once` or `fn_chain` functions. Note, parallel processing is best used when your code block takes more than a minute to run,
+#'  shorter run times will not benefit from parallel processing.
+#'
+#' @param df a data frame containing a water class column, which has already been computed using
+#' \code{\link{define_water_chain}}. The df may include a column named for the coagulant being dosed,
+#' and a column named for the set of coefficients to use.
+#' @param input_water name of the column of Water class data to be used as the input for this function. Default is "defined_water".
+#' @param alum Hydrated aluminum sulfate Al2(SO4)3*14H2O + 6HCO3 -> 2Al(OH)3(am) +3SO4 + 14H2O + 6CO2
+#' @param fecl3 Ferric Chloride FeCl3 + 3HCO3 -> Fe(OH)3(am) + 3Cl + 3CO2
+#' @param fe2so43 Ferric sulfate Fe2(SO4)3 + 6HCO3 -> 2Fe(OH)3(am) +3SO4 + 6CO2
+#' @param coeff String specifying the Edwards coefficients to be used from "Alum", "Ferric", "General Alum", "General Ferric", or "Low DOC" or
+#' named vector of coefficients, which must include: k1, k2, x1, x2, x3, b
+#'
+#' @seealso \code{\link{chemdose_toc}}
+#'
+#' @examples
+#'
+#' library(purrr)
+#' library(furrr)
+#' library(tidyr)
+#' library(dplyr)
+#'
+#' example_df <- water_df %>%
+#'   define_water_chain() %>%
+#'   balance_ions_chain() %>%
+#'   chemdose_ph_chain(alum = 30) %>%
+#'   chemdose_toc_once(input_water = "dosed_chem_water")
+#'
+#' example_df <- water_df %>%
+#'   define_water_chain() %>%
+#'   balance_ions_chain() %>%
+#'   mutate(fecl3 = seq(1, 12, 1),
+#'     coeff = "Ferric") %>%
+#'   chemdose_toc_once(input_water = "balanced_water")
+#'
+#' example_df <- water_df %>%
+#'   define_water_chain() %>%
+#'   balance_ions_chain() %>%
+#'   chemdose_toc_once(input_water = "balanced_water", alum = 40, coeff = "General Alum")
+#'
+#' # Initialize parallel processing
+#' plan(multisession)
+#' example_df <- water_df %>%
+#'   define_water_chain() %>%
+#'   balance_ions_chain() %>%
+#'   mutate(fecl3 = seq(1, 12, 1)) %>%
+#'   chemdose_toc_once(input_water = "balanced_water", coeff = "Ferric")
+#'
+#' # Optional: explicitly close multisession processing
+#' plan(sequential)
+#'
+#' @export
+
+chemdose_toc_once <- function(df, input_water = "defined_water",
+                              alum = 0, fecl3 = 0, fe2so43 = 0, coeff = "Alum") {
+
+  output <- df %>%
+    chemdose_toc_chain(input_water = input_water, output_water = "dosed_chem_water",
+      alum, fecl3, fe2so43, coeff) %>%
+    mutate(dose_chem = furrr::future_map(dosed_chem_water, convert_water)) %>%
+    unnest(dose_chem) %>%
+    select(-dosed_chem_water)
+}
+
+#' Apply `chemdose_toc` within a dataframe and output a column of `water` class to be chained to other tidywater functions
+#'
+#' This function allows \code{\link{chemdose_toc}} to be added to a piped data frame.
+#' Its output is a `water` class, and can therefore be used with "downstream" tidywater functions.
+#' TOC, DOC, and UV254 will be updated based on input chemical doses.
+#'
+#' The data input comes from a `water` class column, as initialized in \code{\link{define_water}} or \code{\link{balance_ions}}.
+#'
+#' If the input data frame has a coagulant(s) name matching a valid coagulant(s), the function will dose that coagulant(s). Note:
+#' The function can only dose a coagulant either a column or from the function arguments, not both.
+#'
+#' The column names must match the chemical names as displayed in \code{\link{chemdose_toc}}.
+#' To see which chemicals can be passed into the function, see \code{\link{chemdose_toc}}.
+#'
+#'  For large datasets, using `fn_once` or `fn_chain` may take many minutes to run. These types of functions use the furrr package
+#'  for the option to use parallel processing and speed things up. To initialize parallel processing, use
+#'  `plan(multisession)` or `plan(multicore)` (depending on your operating system) prior to your piped code with the
+#'  `fn_once` or `fn_chain` functions. Note, parallel processing is best used when your code block takes more than a minute to run,
+#'  shorter run times will not benefit from parallel processing.
+#'
+#' @param df a data frame containing a water class column, which has already been computed using
+#' \code{\link{define_water_chain}}. The df may include a column named for the coagulant being dosed,
+#' and a column named for the set of coefficients to use.
+#' @param input_water name of the column of Water class data to be used as the input for this function. Default is "defined_water".
+#' @param output_water name of the output column storing updated parameters with the class, Water. Default is "coagulated_water".
+#' @param alum Hydrated aluminum sulfate Al2(SO4)3*14H2O + 6HCO3 -> 2Al(OH)3(am) +3SO4 + 14H2O + 6CO2
+#' @param fecl3 Ferric Chloride FeCl3 + 3HCO3 -> Fe(OH)3(am) + 3Cl + 3CO2
+#' @param fe2so43 Ferric sulfate Fe2(SO4)3 + 6HCO3 -> 2Fe(OH)3(am) +3SO4 + 6CO2
+#' @param coeff String specifying the Edwards coefficients to be used from "Alum", "Ferric", "General Alum", "General Ferric", or "Low DOC" or
+#' named vector of coefficients, which must include: k1, k2, x1, x2, x3, b
+#'
+#' @seealso \code{\link{chemdose_toc}}
+#'
+#' @examples
+#'
+#' library(purrr)
+#' library(furrr)
+#' library(tidyr)
+#' library(dplyr)
+#'
+#' example_df <- water_df %>%
+#'   define_water_chain() %>%
+#'   balance_ions_chain() %>%
+#'   chemdose_ph_chain(alum = 30) %>%
+#'   chemdose_toc_chain(input_water = "dosed_chem_water")
+#'
+#' example_df <- water_df %>%
+#'   define_water_chain() %>%
+#'   balance_ions_chain() %>%
+#'   mutate(fecl3 = seq(1, 12, 1),
+#'     coeff = "Ferric") %>%
+#'   chemdose_toc_chain(input_water = "balanced_water")
+#'
+#' example_df <- water_df %>%
+#'   define_water_chain() %>%
+#'   balance_ions_chain() %>%
+#'   chemdose_toc_chain(input_water = "balanced_water", alum = 40, coeff = "General Alum")
+#'
+#' # Initialize parallel processing
+#' plan(multisession)
+#' example_df <- water_df %>%
+#'   define_water_chain() %>%
+#'   balance_ions_chain() %>%
+#'   mutate(fecl3 = seq(1, 12, 1)) %>%
+#'   chemdose_toc_chain(input_water = "balanced_water", coeff = "Ferric")
+#'
+#' # Optional: explicitly close multisession processing
+#' plan(sequential)
+#'
+#' @export
+
+chemdose_toc_chain <- function(df, input_water = "defined_water", output_water = "coagulated_water",
+                               alum = 0, fecl3 = 0, fe2so43 = 0, coeff = "Alum") {
+
+  dosable_chems <- tibble(alum, fecl3, fe2so43)
+
+  chem_inputs_arg <- dosable_chems %>%
+    select_if(~ any(. > 0))
+
+  chem_inputs_col <- df %>%
+    subset(select = names(df) %in% names(dosable_chems)) %>%
+    # add row number for joining
+    mutate(ID = row_number())
+
+
+  if (length(chem_inputs_col) - 1 == 0 & length(chem_inputs_arg) == 0) {
+    warning("No chemical dose found. Create dose column, enter a dose argument, or check availbility of chemical in the chemdose_ph function.")}
+
+  if (length(chem_inputs_col) > 1 & length(chem_inputs_arg) > 0) {
+    stop("Coagulants were dosed as both a function argument and a data frame column. Choose one input method.")}
+  if (length(chem_inputs_col) > 2 | length(chem_inputs_arg) > 1) {
+    stop("Multiple coagulants dosed. Choose one coagulant.")}
+
+  if (length(df$coeff) > 0) {
+    coeff <- tibble(coeff = df$coeff) %>%
+      mutate(ID = row_number())
+  } else {
+    coeff <- tibble(coeff = coeff) %>%
+      mutate(ID = row_number())
+  }
+
+  chem_doses <- chem_inputs_col %>%
+    cross_join(chem_inputs_arg)
+  # Add missing chemical columns
+  chem2 <- dosable_chems %>%
+    subset(select = !names(dosable_chems) %in% names(chem_doses)) %>%
+    cross_join(chem_doses) %>%
+    left_join(coeff, by = "ID") %>%
+    fill(coeff, .direction = "updown")
+
+  output <- df %>%
+    subset(select = !names(df) %in% c("alum", "fecl3", "fe2so43", "coeff")) %>%
+    mutate(ID = row_number()) %>%
+    left_join(chem2, by = "ID") %>%
+    select(-ID) %>%
+    mutate(!!output_water := furrr::future_pmap(list(water = !!as.name(input_water),
+      alum = alum,
+      fecl3 = fecl3,
+      fe2so43 = fe2so43,
+      coeff = coeff),
+    chemdose_toc)) %>%
+    select(!any_of(names(dosable_chems)), any_of(names(chem_doses)))
 }
 
 #' Apply `calculate_corrosion` to a dataframe and create new columns with up to 6 corrosion indices
@@ -1164,6 +1362,10 @@ calculate_corrosion_once <- function(df, input_water = "defined_water", index = 
 calculate_corrosion_chain <- function(df, input_water = "defined_water", output_water = "corrosion_indices",
                                       index = c("aggressive", "ryznar", "langelier", "ccpp", "larsonskold", "csmr"),
                                       form = "calcite") {
+
+  if (any(!index %in% c("aggressive", "ryznar", "langelier", "ccpp", "larsonskold", "csmr"))) {
+    stop("Index must be one or more of c('aggressive', 'ryznar', 'langelier', 'ccpp', 'larsonskold', 'csmr')") }
+
   index = list(index)
 
   output <- df %>%