+++ tests/testthat/test-corrosion_indices.R
@@ -1,28 +1,51 @@
 test_that("most indices won't work without ca, cl, so4", {
 
-  water <- suppressWarnings(define_water(ph = 8, temp = 25, alk = 200, tds = 238)) %>%
-    calculate_corrosion()
+  water <- suppressWarnings(define_water(ph = 8, temp = 25, alk = 200, tds = 238))
 
 
-  expect_equal(water@aggressive, NA_real_)
-  expect_equal(water@ryznar, NA_real_)
-  expect_equal(water@langelier, NA_real_)
-  expect_equal(round(water@ccpp), -9)
-  expect_equal(water@larsonskold, 0)
-  expect_equal(water@csmr, NA_real_)
+
+  expect_equal(suppressWarnings(calculate_corrosion(water, index = "aggressive"))@aggressive, NA_real_)
+  expect_error(suppressWarnings(calculate_corrosion(water, index = "ryznar")))
+  expect_error(suppressWarnings(calculate_corrosion(water, index = "langelier")))
+  expect_error(suppressWarnings(calculate_corrosion(water, index = "ccpp")))
+  expect_equal(suppressWarnings(calculate_corrosion(water, index = "larsonskold"))@larsonskold, NA_real_)
+  expect_equal(suppressWarnings(calculate_corrosion(water, index = "csmr"))@csmr, NA_real_)
+
+})
+
+test_that("function catches index typos", {
+
+  water <- suppressWarnings(define_water(ph = 8, temp = 25, alk = 200, tds = 238,
+    tot_hard = 100, cl = 40, so4 = 40))
+
+  expect_error(calculate_corrosion(water, index = "csr"))
+  expect_error(calculate_corrosion(water, index = c("aggressive", "ccccp")))
+  expect_error(calculate_corrosion(water, index = c("ai", "ryznar", "ccpp", "csmr", "langelier")))
+  expect_no_error(calculate_corrosion(water, index = c("ryznar", "csmr", "larsonskold"))) # no error
+})
+
+test_that("warnings are present when parameters used in calculations are estimated by tidywater.", {
+
+  water1 <- suppressWarnings(define_water(8, 25, 200, 200))
+  water2 <- suppressWarnings(define_water(8, 25, 200, 200, na = 100, cl = 100)) %>% balance_ions()
+
+  expect_warning(calculate_corrosion(water1, index = "aggressive"))
+  expect_warning(calculate_corrosion(water2, index = "csmr"))
+  expect_warning(calculate_corrosion(water2, index = "larsonskold"))
 
 })
 
 test_that("aggressive index works", {
+  suppressWarnings({
+    water1 <- define_water(ph = 8, temp = 25, alk = 200, ca_hard = 200) %>%
+      calculate_corrosion(index = "aggressive")
 
-  water1 <- suppressWarnings(define_water(ph = 8, temp = 25, alk = 200, ca_hard = 200)) %>%
-    calculate_corrosion(index = "aggressive")
+    water2 <- define_water(ph = 8, temp = 25, alk = 15, ca_hard = 200) %>%
+      calculate_corrosion(index = "aggressive")
 
-  water2 <- suppressWarnings(define_water(ph = 8, temp = 25, alk = 15, ca_hard = 200)) %>%
-    calculate_corrosion(index = "aggressive")
-
-  water3 <- suppressWarnings(define_water(ph = 8, temp = 25, alk = 15, tot_hard = 150)) %>%
-    calculate_corrosion(index = "aggressive")
+    water3 <- define_water(ph = 8, temp = 25, alk = 15, tot_hard = 150) %>%
+      calculate_corrosion(index = "aggressive")
+  })
 
   expect_equal(round(water1@aggressive), 13) # high alk
   expect_equal(round(water2@aggressive), 11) # low alk
@@ -31,16 +54,17 @@ test_that("aggressive index works", {
 })
 
 test_that("csmr works", {
+  suppressWarnings({
+    water1 <- define_water(ph = 8, temp = 25, alk = 200, cl = 100, so4 = 1) %>%
+      calculate_corrosion(index = "csmr")
 
-  water1 <- suppressWarnings(define_water(ph = 8, temp = 25, alk = 200, cl = 100, so4 = 1)) %>%
-    calculate_corrosion(index = "csmr")
+    water2 <- define_water(ph = 8, temp = 25, cl = 2, so4 = 150) %>%
+      calculate_corrosion(index = "csmr")
 
-  water2 <- suppressWarnings(define_water(ph = 8, temp = 25, cl = 2, so4 = 150)) %>%
-    calculate_corrosion(index = "csmr")
-
-  water3 <- suppressWarnings(define_water(ph = 8, temp = 25, alk = 15, tot_hard = 150, so4 = 5)) %>%
-    balance_ions() %>%
-    calculate_corrosion(index = "csmr")
+    water3 <- define_water(ph = 8, temp = 25, alk = 15, tot_hard = 150, so4 = 5) %>%
+      balance_ions() %>%
+      calculate_corrosion(index = "csmr")
+  })
 
   expect_equal(round(water1@csmr), 100) # high cl, low so4
   expect_equal(round(water2@csmr, 2), 0.01) # low cl high so4
@@ -55,7 +79,7 @@ test_that("larsonskold works", {
   water2 <- suppressWarnings(define_water(ph = 8, temp = 25, alk = 200, cl = 2, so4 = 150)) %>%
     calculate_corrosion(index = "larsonskold")
 
-  water3 <- suppressWarnings(define_water(ph = 8, temp = 25, alk = 200, tot_hard = 150, so4 = 5)) %>%
+  water3 <- suppressWarnings(define_water(ph = 8, temp = 25, alk = 200, tot_hard = 150, cl = 50, so4 = 30)) %>%
     balance_ions() %>%
     calculate_corrosion(index = "larsonskold")
 
@@ -64,8 +88,15 @@ test_that("larsonskold works", {
 
   expect_equal(round(water1@larsonskold, 1), 0.7) # high cl, low so4
   expect_equal(round(water2@larsonskold, 1), 0.8) # low cl high so4
-  expect_equal(round(water3@larsonskold, 2), 0.03) # use balance ions to get chloride
+  expect_equal(round(water3@larsonskold, 2), 0.51) # use balance ions to get chloride
   expect_equal(round(water4@larsonskold), 74) # low alk
+})
+
+test_that("Corrosion index calculations work when IS is NA.", {
+  water1 <- suppressWarnings(define_water(ph = 8, temp = 25, alk = 200, ca_hard = 200))
+
+  expect_no_error(calculate_corrosion(water1, index = c("langelier", "ryznar", "ccpp")))
+
 })
 
 # test answers will probably change as we figure out which ph_s to use. For now, I'm using MWH's ph_s.