+++ tests/testthat/test-general_functions.R
@@ -17,10 +17,30 @@ test_that("Define water calculates correct carbonate balance.", {
   })
 
   expect_equal(water1@ph, 7)
-  expect_equal(round(water1@tot_co3, 5), 0.00244)
-  expect_equal(round(water1@hco3, 3), 0.002)
+  expect_equal(signif(water1@tot_co3, 2), 0.0024)
+  expect_equal(signif(water1@hco3, 2), 0.002)
 
 })
+
+test_that("Define water calculates correct TDS/IS/cond.", {
+  water1 <- suppressWarnings(define_water(ph = 7, temp = 25, alk = 100, tds = 200))
+  water2 <- suppressWarnings(define_water(ph = 7, temp = 25, alk = 100, cond = 312))
+  water3 <- suppressWarnings(define_water(ph = 7, temp = 25, alk = 100, tot_hard = 100, cl = 100, so4 = 30))
+
+  expect_true(grepl("cond", water1@estimated))
+  expect_true(grepl("tds", water2@estimated))
+  expect_true(grepl("cond", water3@estimated) & grepl("tds", water3@estimated))
+
+  expect_equal(round(water1@cond), 312)
+  expect_equal(round(water2@tds), 200)
+  expect_equal(signif(water1@is, 2), .005)
+  expect_equal(signif(water2@is, 2), .005)
+  expect_equal(signif(water3@is, 2), .005)
+  expect_equal(round(water3@cond), 315)
+  expect_equal(round(water3@tds), 201)
+
+})
+
 
 test_that("Define water gives missing value warnings.", {
   expect_warning(define_water(alk = 100, temp = 20, tot_hard = 50, ca_hard = 50, na = 10, k = 10, cl = 10, so4 = 10, tds = 100,
@@ -29,13 +49,6 @@ test_that("Define water gives missing value warnings.", {
   expect_warning(define_water(ph = 7, temp = 20, tot_hard = 50, ca_hard = 50, na = 10, k = 10, cl = 10, so4 = 10, tds = 100,
     doc = 5, toc = 5, uv254 = .1, br = 50),
   "Missing.+alkalinity.+")
-  expect_warning(define_water(ph = 7, alk = 100, temp = 20, tot_hard = 50, ca_hard = 50, tds = 100,
-    doc = 5, toc = 5, uv254 = .1, br = 50),
-  "Missing.+cations.+")
-  expect_warning(define_water(ph = 7, alk = 100, temp = 20, tot_hard = 50, ca_hard = 50, na = 0, k = 0, cl = 0, so4 = 0,
-    doc = 5, toc = 5, uv254 = .1, br = 50),
-  "Major ions missing.+")
-
   expect_warning(define_water(ph = 7, alk = 100, temp = 20, tot_hard = 50, ca_hard = 50, na = 10, k = 10, cl = 10, so4 = 10,
     toc = 5, uv254 = .1, br = 50),
   "Missing.+DOC+")
@@ -64,6 +77,25 @@ test_that("define_water handles organics inputs correctly.", {
   expect_equal(water1@doc, 3.325)
   expect_equal(round(water2@toc, 3), 3.684)
   expect_equal(water3@uv254, NA_real_)
+})
+
+test_that("define_water correctly specifies when estimates are used.", {
+
+  water1 <- suppressWarnings(define_water(ph = 7, temp = 25, alk = 100, tot_hard = 50, na = 100, cl = 100))
+  water2 <- suppressWarnings(define_water(ph = 7, toc = 3.5, uv254 = 0.1))
+  water3 <- suppressWarnings(define_water(ph = 7, temp = 25, alk = 100, tot_hard = 50, na = 100, cl = 100, toc = 3.5, doc = 3.5,
+    ca = 40, tds = 100))
+
+  expect_true(grepl("tds", water1@estimated))
+  expect_true(grepl("cond", water1@estimated))
+  expect_true(grepl("ca", water1@estimated))
+  expect_true(grepl("doc", water2@estimated))
+
+  expect_false(grepl("tds", water3@estimated))
+  expect_false(grepl("ca", water3@estimated))
+  expect_false(grepl("doc", water3@estimated))
+  expect_true(grepl("cond", water3@estimated))
+
 })
 
 
@@ -188,6 +220,25 @@ test_that("Balance ions results in neutral charge.", {
 
 })
 
+test_that("Balance ions only updates TDS/cond/IS when appropriate.", {
+  water1 <- suppressWarnings(define_water(ph = 7, temp = 25, alk = 100, tds = 100))
+  water2 <- balance_ions(water1)
+  water3 <- suppressWarnings(define_water(ph = 7, temp = 25, alk = 100, cond = 100))
+  water4 <- balance_ions(water3)
+  water5 <- suppressWarnings(define_water(ph = 7, temp = 25, alk = 100, na = 100, tot_hard = 100, cl = 100, so4 = 100))
+  water6 <- balance_ions(water5)
+
+  expect_false(grepl("tds", water2@estimated))
+  expect_equal(round(water1@tds), round(water2@tds))
+  expect_false(grepl("cond", water4@estimated))
+  expect_equal(round(water3@tds), round(water4@tds))
+  expect_true(grepl("cond", water5@estimated) & grepl("tds", water5@estimated))
+  expect_true(grepl("cond", water6@estimated) & grepl("tds", water6@estimated))
+  expect_error(expect_equal(round(water5@tds), round(water6@tds)))
+  expect_error(expect_equal(signif(water5@is, 2), signif(water6@is, 2)))
+
+})
+
 
 # Calculate alpha carbonate ----
 
@@ -230,7 +281,7 @@ test_that("Ionic strength calc in define water works.", {
   is_calced <- 0.5 * ((water@na + water@cl + water@k + water@hco3 + water@h2po4 + water@h + water@oh + water@tot_ocl) * 1^2 +
     (water@ca + water@mg + water@so4 + water@co3 + water@hpo4) * 2^2 +
     (water@po4) * 3^2)
-  expect_equal(water@is, is_calced)
+  expect_equal(signif(water@is, 3), signif(is_calced, 3))
 
 })